codal: began removal of idle and system components
The concept of idle components and system components was placed in the
micro:bit runtime from an early stage, this solution is not scalable
and violates some principles of the runtime.

As we move away from the reliance upon function callbacks from the
timer, and increase the reliance on a core codal concept, events, it
makes sense to instead move idle components to rely on events.