codal: Added CMake Target name definitions for any libs requiring linking to the target.
I engineered a poor solution whereby any target was cloned as "codal-target" so that any libs requiring linking to the target could use a static name. This seemed like a good solution initially, on reflection it was a bit ugly.

Target libs are now cloned like any other library under the libary name, and a definition has been added to CMake to indicate the name of the codal-target