Add support for asynchronous DataStreams
DataStream now supports an asynchronous mode. In this mode, pullRequests
result in the buffer being extracted form the source and held in the
DataStream, but downstream processing is deferred until scheduled later. This
is particualrly useful for decoupling data streams generated in interrupt
context.

The default mode remains synchronous, and behaves as per previous versions.